Switch 37-fix_IndexError_list_index_out_of_range.patch to default to first item
The ArchLinux patch made the wired profile to default to "Add new
profile" which may be confusing. A different patch from one of the
upstream bug reports is slightly simpler and defaults to the first
profile which makes more sense to me. So switching the patch to the
one posted in https://bugs.launchpad.net/wicd/+bug/1448685/comments/1